+++ server/controllers/api/claim/get/addGetResultsToFileData.js
@@ -1,0 +1,5 @@
+module.exports = (fileInfo, getResult) => {
+  fileInfo.fileName = getResult.file_name;
+  fileInfo.filePath = getResult.download_path;
+  return fileInfo;
+};
