+++ server/controllers/api/claim/get/createFileData.js
@@ -1,0 +1,13 @@
+module.exports = ({ name, claimId, outpoint, height, address, nsfw, contentType }) => {
+  return {
+    name,
+    claimId,
+    outpoint,
+    height,
+    address,
+    fileName: '',
+    filePath: '',
+    fileType: contentType,
+    nsfw,
+  };
+};
